converting base-creator to vue component mixin
